feature/add_commchannels_strat
add_commchannels_strat
