+++ app/repository/commchannel_repository.py
@@ -21,7 +21,6 @@ class CommChannelRepository:
     def _get_user_commchannels(tx, name):
         nodes = tx.run("MATCH (u:User {name:$name})-[:HAVE]->(a)-[:USE]->(cc:ChannelCommunication) RETURN collect(distinct cc)", name=name)
         results = [record for record in nodes.data()]
-        print(results[0]['collect(distinct cc)'])
         return results[0]['collect(distinct cc)']
 
     def get_user_commchannels(self, name):
@@ -29,11 +28,3 @@ class CommChannelRepository:
             commchannel = session.execute_read(self._get_user_commchannels, name)
             self.close()
             return commchannel
-
-def serialize_commchannel(commchannel):
-    serialize_dict, ret = {}, {}
-    for i, cc in enumerate(commchannel):
-        serialize_dict[f"cc{i}"] = cc["type"]
-    ret["cc"] = serialize_dict
-    print(ret)
-    return commchannel
